refactor: simplify map message file name
